pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "URAG_V2"
version = "0.1.0"
description = "Use PydanticAI"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "nltk>=3.9.1",
  "pydantic-ai-slim[google,mistral]",
  "dotenv>=0.9.9",
  "google>=3.0.0",
  "google-generativeai>=0.8.5",
  "python-dotenv>=1.1.1",
  "pymilvus>=2.6.2",
  "pytest>=8.4.2",
  "sentence-transformers>=5.1.0",
  "langchain-text-splitters>=0.3.11",
  "langchain-community>=0.3.29",
  "langchain-experimental>=0.3.4",
]

[project.optional-dependencies]
parse = [
  "beautifulsoup4>=4.12.3",
  "pdfminer.six>=20221105",
  "PyMuPDF>=1.24.9",
  "ftfy>=6.2.0",
  "rapidfuzz>=3.9.6",
]

llm = [
  "openai>=1.40.0",
  "anthropic>=0.34.0",
]

dev = [
  "ruff>=0.5.7",
  "black>=24.8.0",
  "mypy>=1.11.1",
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "types-PyYAML>=6.0.12.20240808",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src", "data", "config", "examples", "notebooks" ]
include = ["*"]  


[tool.setuptools.package-data]
"*" = ["**/*.yaml", "**/*.yml", "**/*.txt", "**/*.json"]

# Ruff (lint) + Black (format) gợi ý
[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP"]
extend-exclude = ["data/", "notebooks/"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.pyright]
include = ["src", "examples", "tests"]
extraPaths = ["src", "src/URag_F"]
venv = ".venv"
venvPath = "."
